Move ApiUser to auth0
